backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-multipart
python-jose
passlib==1.7.4
bcrypt==4.0.1